Do not merge unknown values when returning raw data
We noticed in the zf-apigility-admin when submitting service fields that
values were being duplicated (see zfcampus/zf-apigility-admin#345). The
reason is because, by default, raw data is used, but a change for the
1.3 release series modified the logic of the content validation listener
to no longer return raw data immediately, but to also test that it does
not contain unknown fields. In the default use case, when unknown fields
are allowed, the unknown fields were being merged back into the data
set, causing duplication.

This patch adds a conditional to check if raw data is being used after
the check to see if unknown fields are allowed, and, if so, passes the
data to the parameter data container and returns immediately, instead of
merging.