Protect event state
We were cloning, but not using the cloned event; this makes sense, as a listener
may push something back into it to use later. However, we need to ensure that
later listeners on the same event that triggered this listener get the same
event name. As such, this memoizes the original name, and then resets it after
triggering the validation listeners.